refactor: remove import side effect from session.server.ts
