+++ app/trpc/context.ts
@@ -3,7 +3,7 @@ import type { Session } from "@remix-run/server-runtime";
 import type { FetchCreateContextFnOptions } from "@trpc/server/adapters/fetch";
 import type { TT } from "../db/drizzle-client.server";
 import { getSessionUser } from "../utils/auth";
-import { getRequestSession } from "../utils/session-utils";
+import { getRequestSession } from "../utils/session.server";
 import { middlewareFactory } from "./factory";
 
 export type TrpcAppContext = {
