+++ app/utils/session.server.ts
@@ -1,14 +1,38 @@
-import { createCookieSessionStorage } from "@remix-run/node";
-import { initializeConfigServer, serverConfig } from "./config";
+import { once } from "@hiogawa/utils";
+import {
+  Session,
+  SessionStorage,
+  createCookieSessionStorage,
+} from "@remix-run/node";
+import { serverConfig } from "./config";
 
-// TODO: avoid config dependent side-effect
-initializeConfigServer();
+export let sessionStore: SessionStorage;
 
-const { getSession, commitSession } = createCookieSessionStorage({
-  cookie: {
-    httpOnly: true,
-    secrets: [serverConfig.APP_SESSION_SECRET],
-  },
+export const initializeSessionStorage = once(() => {
+  sessionStore = createCookieSessionStorage({
+    cookie: {
+      httpOnly: true,
+      secrets: [serverConfig.APP_SESSION_SECRET],
+    },
+  });
 });
 
-export { getSession, commitSession };
+//
+// utils
+//
+
+export async function getRequestSession(request: Request): Promise<Session> {
+  return sessionStore.getSession(request.headers.get("cookie"));
+}
+
+export async function getResponseSession(response: Response): Promise<Session> {
+  return sessionStore.getSession(response.headers.get("set-cookie"));
+}
+
+export async function withResponseSession(
+  response: Response,
+  session: Session
+): Promise<Response> {
+  response.headers.set("set-cookie", await sessionStore.commitSession(session));
+  return response;
+}
